requirements.txt
Flask==2.3.2
rembg==2.0.17
Pillow==9.0.0
